Move some helper functions to utils
